refactor(config): move lenient-parsing logic into src/config/lenient.rs
src/config.rs had grown to ~900 lines; splitting out the field-by-field
lenient extraction helpers keeps the core RawConfig/AppConfig definitions
separate from the parsing-recovery logic.
